pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["paul-leonard <paul.ryan.leonard@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^3.8.4"
Django = "^3.1.4"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
